Solve Day 6
This one some logic bled into the actual file loading private method,
but overall these solutions are ok. I hate initializing a sum outside of
a loop since that is an immediate tell I am not doing the "rubyish" way
of doing it, so that's one I should be revisiting.