Complete Day 7
This one is tough and since I chose to store my values in a list and not
backfill the data of the sub-objects(HeldBags as I called them), it made
part 1 of this laughable inefficent. I think the best bet here is a
hash/tree where each line you check existing places and backfill until
you have the whole structure built. My way I have to query my master
list for every single HeldBag which is brutal. Part 2 is quick, but part
1 seems to take like 10 seconds.... yikes.