Solve Day 8
This was a fun one, My answer isn't quite as efficient as I would like,
but part 1 i think is quite nice. This one was really helpful to parse
into objects as it made it really easy to access everything throughout.

The hardest part of this was in part 2 getting it to make a deep copy of
the list for each test attempt if it was infinite. The answer I settled
on, is it is easier to just re-parse the whole damn list than to
actually make a deep copy in ruby, so that was what I used for this
case.