Solve part 2
Ok, please don't judge me on this one. Clearly the ruby-ish way here is
to pass a block to the method that does the -1 or +1 or whatever based
on which direction, but its late and im going to bed and copy/pasting is
an easy way to get the answer without having to really think at all. Yea
I know this is laughably bad.